Compute Built-Using with go list, not Build-Depends
Built-Using for a binary is meant to include all packages that are included in
the binary itself, but using Build-Depends only pulls in the direct
dependencies. Use go list and dpkg-query --search instead to find out which
(debian) packages installed the (go) packages that were actually used during
the build instead.

As a bonus, this is shorter and arguably simpler that what it replaces.